requirements.txt
asgiref==3.4.1
Django==4.0
django-mptt==0.13.4
django-cors-headers==3.10.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
